+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - Improved README files with description of endpoints and how to test it
 ### Fixed
 - Parsing of CaseData csv with more than one individuals associated to the same variant
+- dry-run enpoint, to return success when ClinVar returns a 204 successful response